requirements.txt
aiohttp==3.9.5
aiohttp-sse-client==0.2.1
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
bitarray==2.9.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
pycparser==2.22
pycryptodomex==3.20.0
pymongo==4.7.3
PyNaCl==1.5.0
pypng==0.20220715.0
python-dotenv==1.0.1
python-telegram-bot==21.2
pytonconnect==0.3.1
pytoniq-core==0.1.36
qrcode==7.4.2
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.0
urllib3==2.2.1
x25519==0.0.2
yarl==1.9.4
